requirements.txt
Flask==3.0.3
flask_wtf==1.2.1
pymongo==4.8.0
python-dotenv==1.0.1
WTForms==3.1.2
